# Optimized matrix-rendering