--- a/src/matrix_viz/core.clj
+++ b/src/matrix_viz/core.clj
@@ -38,7 +38,7 @@
     (m/eseq data)))
 
 (defn- mget
-  ([data x0] data)
+  ([data] data)
   ([data x0]
    (if (t/tensor? data)
      (t/mget data x0)