From my measurements, these 2 optimizations increase by 4x the efficiency of BufferedImage-generation in `matrix-viz.core/save-matrix-as-png`.

---
## Performance measurements

I measured on my laptop the performance of generating the BufferedImage of a 1600x1600 GridFire `:burn-time-matrix`. The evaluated expression looked like:

```clojure
(matrix-viz.core/matrix-to-buffered-image
              :color 1 -1.0
              (:burn-time-matrix matrices))
```

This 1st flame graph, made before the optimizations, shows that these were significant sources of overhead:
![01-cpu-flamegraph-2022-07-01-18-12-33](https://user-images.githubusercontent.com/5859120/177003095-98409105-9c3a-45d2-bddc-fa3cf0099644.svg)

Before optimizing, criterium-benchmarking yielded:

```
Evaluation count : 120 in 60 samples of 2 calls.
             Execution time mean : 1.006235 sec
    Execution time std-deviation : 30.846738 ms
   Execution time lower quantile : 984.665431 ms ( 2.5%)
   Execution time upper quantile : 1.074640 sec (97.5%)
                   Overhead used : 1.887161 ns
```

... and after optimizing:
```
Evaluation count : 240 in 60 samples of 4 calls.
             Execution time mean : 266.026682 ms
    Execution time std-deviation : 50.351484 ms
   Execution time lower quantile : 244.967024 ms ( 2.5%)
   Execution time upper quantile : 334.066427 ms (97.5%)
                   Overhead used : 1.884948 ns
```
 
Runtime details:

```
{:spec-vendor            "Oracle Corporation",
 :spec-name              "Java Virtual Machine Specification",
 :vm-version             "25.282-b08",
 :name                   "49196@robo",
 :clojure-version-string "1.10.3",
 :java-runtime-version   "1.8.0_282-b08",
 :java-version           "1.8.0_282",
 :vm-name                "OpenJDK 64-Bit Server VM",
 :vm-vendor              "AdoptOpenJDK",
 :clojure-version        {:major 1, :minor 10, :incremental 3, :qualifier nil},
 :spec-version           "1.8",
 :sun-arch-data-model    "64",
 :input-arguments        ["-Dclojure.libfile=/private/var/folders/62/3v4gckg15t706bjzkjw155n00000gq/T/libfile5.libs"
                          "-javaagent:/Applications/IntelliJ IDEA CE.app/Contents/lib/idea_rt.jar=61033:/Applications/IntelliJ IDEA CE.app/Contents/bin"
                          "-Dfile.encoding=UTF-8"]}
```

For completeness, here is the after-optimizing flame graph:
![02-cpu-flamegraph-2022-07-02-15-28-20](https://user-images.githubusercontent.com/5859120/177003436-4111268d-46e2-4dc4-803f-cff294a778cc.svg)


... and here is the PNG-encoded image:
![my-test-burn-matrix](https://user-images.githubusercontent.com/5859120/177003451-44239098-4997-4ad2-bb90-372dd3ec6080.png)